+++ src/pnfl_gameplan/__init__.py
@@ -1,5 +1,16 @@
 """Library applying PNFL league rules and validation to a Front Page Sports Football Pro '98 gameplan (.pln)."""
 
+from fbpro98_gameplan import (
+    CustomPlay,
+    InvalidGamePlanError,
+    Play,
+    ProfileType,
+    StockPlay,
+    parse_gameplan,
+    read_gameplan,
+    write_gameplan,
+)
+
 from pnfl_gameplan.model import (
     PnflGamePlan,
     PnflRuleWarning,
@@ -16,12 +27,20 @@ from pnfl_gameplan.validators import validate_gameplan
 
 __all__ = [
     "PNFL_RULES",
+    "CustomPlay",
     "DefenseCategoryRule",
+    "InvalidGamePlanError",
     "OffenseCategoryRule",
+    "Play",
     "PnflGamePlan",
     "PnflRuleWarning",
     "PnflRules",
+    "ProfileType",
     "RuleName",
+    "StockPlay",
     "Violation",
+    "parse_gameplan",
+    "read_gameplan",
     "validate_gameplan",
+    "write_gameplan",
 ]